Rename join_cluster to fit sender/recvr pattern that is unfolding
