+++ lib/swim/join-sender.js
@@ -19,11 +19,11 @@
 // THE SOFTWARE.
 'use strict';
 
-var captureHost = require('./util.js').captureHost;
-var errors = require('./errors.js');
-var isEmptyArray = require('./util.js').isEmptyArray;
-var numOrDefault = require('./util.js').numOrDefault;
-var safeParse = require('./util.js').safeParse;
+var captureHost = require('../util.js').captureHost;
+var errors = require('../errors.js');
+var isEmptyArray = require('../util.js').isEmptyArray;
+var numOrDefault = require('../util.js').numOrDefault;
+var safeParse = require('../util.js').safeParse;
 var TypedError = require('error/typed');
 
 var JoinAbortedError = TypedError({
